+++ src/model/Properties.ts
@@ -1,4 +1,4 @@
-import { PatternColor } from './PatternColor';
+import { Color } from './Color';
 import { PatternTotals } from './PatternTotals';
 
 /**
@@ -6,17 +6,17 @@ import { PatternTotals } from './PatternTotals';
  */
 export class Properties {
     /**
-     * @param patternColors - The list of PatternColors that are used in the pattern.
+     * @param colors - The list of colors that are used in the pattern.
      * @param patternTotals - The PatternTotals containing stitch counts by type and color
      * @param stitchWidth - The width, in stitches, of the pattern.
      * @param stitchHeight - The height, in stitches, of the pattern.
      * @param notes - Any notes or comments about the pattern.
      */
     constructor(
-        public patternColors: PatternColor[],
-        public patternTotals?: PatternTotals,
-        public stitchWidth?: number,
-        public stitchHeight?: number,
+        public colors: Color[],
+        public patternTotals: PatternTotals,
+        public stitchWidth: number,
+        public stitchHeight: number,
         public notes?: string
     ) {}
 }